feat: #30 - 캘린더 근무 Entity에 맞춰 코드 수정, 근무 시간 계산 유틸 메서드 추가

+++ MOUP/MOUP/Domain/Entities/FilterWorkplace.swift
@@ -6,12 +6,12 @@
 //
 
 /// 근무지/매장 필터 Entity
-/// - `workplaceId`: 근무지/매장 ID `Int`
+/// - `workplaceId`: 근무지/매장 ID `Int64`
 /// - `workplaceName`: 근무자/매장 이름 `String`
-///   - 예시: "세븐일레븐 동탄제일점"
+///   - 예시: `"세븐일레븐 동탄제일점"`
 /// - `isShared`: 공유 캘린더 여부 `Bool`
 struct FilterWorkplace {
-    let workplaceId: Int
+    let workplaceId: Int64
     let workplaceName: String
     let isShared: Bool
 }
